feat: userProfile, tutorProfile, timeSlots models

+++ backend/src/main/java/com/slemanski/backend/features/auth/repository/MyUserRepository.java
@@ -3,6 +3,6 @@ package com.slemanski.backend.features.auth.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.slemanski.backend.features.auth.model.MyUser;
 
-public interface MyUserRepository extends JpaRepository<MyUser, Integer> {
+public interface MyUserRepository extends JpaRepository<MyUser, Long> {
     MyUser findByUsername(String username);
 }
